pyproject.toml
[project]
name = "contractguard-ai"
version = "0.1.0"
description = "Enterprise Contract Intelligence Platform - Multi-Agent System with Google ADK"
readme = "README.md"
license = {text = "Apache-2.0"}
requires-python = ">=3.11"
authors = [
    {name = "Tejaswi Mahapatra", email = "tejaswimahapatra208@gmail.com"}
]
keywords = ["ai", "agents", "google-adk", "gemini", "contracts", "rag", "multi-agent"]
classifiers = [
    "Development Status :: 4 - Beta",
    "Intended Audience :: Developers",
    "License :: OSI Approved :: Apache Software License",
    "Programming Language :: Python :: 3.11",
    "Programming Language :: Python :: 3.12",
    "Topic :: Scientific/Engineering :: Artificial Intelligence",
]

dependencies = [
    # Google ADK - Core Agent Framework (with A2A support)
    "google-adk[a2a]>=1.0.0",
    "google-genai>=0.5.0",
    "google-generativeai>=0.8.0",
    "litellm>=1.0.0",  # Required by google-adk for LiteLlm model support

    # FastAPI for API layer
    "fastapi>=0.109.0",
    "uvicorn[standard]>=0.27.0",
    "pydantic>=2.5.0",
    "pydantic-settings>=2.1.0",
    "python-multipart>=0.0.6",
    "websockets>=12.0",

    # Vector Database & Embeddings
    "weaviate-client>=4.4.0",
    "sentence-transformers>=2.3.0",

    # Document Processing
    "pypdf>=3.17.0",
    "pdfplumber>=0.10.0",

    # Async Support
    "aiofiles>=23.2.0",
    "httpx>=0.26.0",
    "redis>=5.0.0",

    # Database
    "sqlalchemy[asyncio]>=2.0.0",
    "asyncpg>=0.29.0",
    "alembic>=1.13.0",

    # Object Storage
    "minio>=7.2.0",

    # Memory
    "mem0ai>=0.1.0",
    "chromadb>=0.4.0",

    # MCP Support
    "mcp>=1.0.0",

    # Observability
    "opentelemetry-api>=1.22.0",
    "opentelemetry-sdk>=1.22.0",
    "opentelemetry-exporter-otlp>=1.22.0",
    "structlog>=24.1.0",

    # Utilities
    "python-dotenv>=1.0.0",
    "tenacity>=8.2.0",
    "rich>=13.7.0",

    # PDF generation (for sample contracts)
    "markdown>=3.5.0",
    "fpdf2>=2.7.0",
]

[project.optional-dependencies]
dev = [
    "pytest>=7.4.0",
    "pytest-asyncio>=0.23.0",
    "pytest-cov>=4.1.0",
    "mypy>=1.8.0",
    "ruff>=0.1.0",
    "pre-commit>=3.6.0",
]

[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[tool.hatch.build.targets.wheel]
packages = ["src"]

[tool.ruff]
target-version = "py311"
line-length = 100
select = ["E", "F", "I", "N", "W", "UP"]
ignore = ["E501"]

[tool.mypy]
python_version = "3.11"
strict = true
ignore_missing_imports = true

[tool.pytest.ini_options]
asyncio_mode = "auto"
testpaths = ["tests"]
